# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ramon api",
        "description": "api for games, movies and books",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/books": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Get all books",
                "description": "Returns all books",
                "operationId": "bookIndex",
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/book/{id}": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Get book by index",
                "description": "Returns a book by the given id",
                "operationId": "showBooks",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Book"
                ],
                "summary": "Updates and returns a book object",
                "description": "Updates and returns a book",
                "operationId": "editBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update a Book object ID not required",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "422": {
                        "description": "Unproccessed data"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Book"
                ],
                "summary": "Delete a book",
                "description": "Delete a book with the given id",
                "operationId": "destroyBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Books"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "No content"
                    }
                }
            }
        },
        "/api/book": {
            "post": {
                "tags": [
                    "Book"
                ],
                "summary": "Creates and returns a book object",
                "description": "Creates and returns a book",
                "operationId": "createBook",
                "requestBody": {
                    "description": "Update a Book object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "422": {
                        "description": "Unproccessed data"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/book/giveTag/{id}": {
            "post": {
                "tags": [
                    "Book"
                ],
                "summary": "add a tag to a book",
                "description": "add tag",
                "operationId": "tagBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Tag object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "422": {
                        "description": "Unproccessed data"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/api/game/{id}": {
            "get": {
                "tags": [
                    "Game"
                ],
                "summary": "Get game by index",
                "description": "Returns a game by the given id",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Game"
                ],
                "summary": "Edit game",
                "description": "Updates a game Consuming a Game object",
                "operationId": "edit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update a Game object ID not required",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "422": {
                        "description": "Unproccessed data"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Game"
                ],
                "summary": "Delete a game",
                "description": "Delete a game with the given id",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/game": {
            "post": {
                "tags": [
                    "Game"
                ],
                "summary": "Creates and returns a game object",
                "description": "Creates and returns a game",
                "operationId": "create",
                "requestBody": {
                    "description": "Create a Game object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "422": {
                        "description": "Unproccessed data"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/game/giveTag/{id}": {
            "post": {
                "tags": [
                    "Game"
                ],
                "summary": "add a tag to a game",
                "description": "add tag",
                "operationId": "tagGame",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Tag object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "422": {
                        "description": "Unproccessed data"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/api/movie/{id}": {
            "get": {
                "tags": [
                    "Movie"
                ],
                "summary": "Get movie by index",
                "description": "Returns a movie by the given id",
                "operationId": "showMovies",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Movie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movie"
                    }
                }
            },
            "put": {
                "tags": [
                    "Movie"
                ],
                "summary": "Updates and returns a movie object",
                "description": "Updates and returns a movie",
                "operationId": "editMovie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "update with a Movie object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "422": {
                        "description": "Unproccessed data"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Movie"
                ],
                "summary": "Delete a movie",
                "description": "Delete a movie with the given id",
                "operationId": "destroyMovie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Movie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/movie": {
            "post": {
                "tags": [
                    "Movie"
                ],
                "summary": "Creates and returns a movie object",
                "description": "Creates and returns a movie",
                "operationId": "createMovie",
                "requestBody": {
                    "description": "update with a Movie object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "422": {
                        "description": "Unproccessed data"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/movie/giveTag/{id}": {
            "post": {
                "tags": [
                    "Movie"
                ],
                "summary": "add a tag to a movie",
                "description": "add tag",
                "operationId": "tagMovie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Tag object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "422": {
                        "description": "Unproccessed data"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/api/tag/{id}": {
            "get": {
                "tags": [
                    "Tag"
                ],
                "summary": "Get tag by index",
                "description": "Returns a tag by the given id",
                "operationId": "showTags",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tag id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tag"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tag"
                ],
                "summary": "Delete a tag",
                "description": "Delete a tag with the given id",
                "operationId": "destroyTag",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tag id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Tag"
                ],
                "summary": "Updates and returns a tag object",
                "description": "Updates and returns a tag",
                "operationId": "editTag",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Tag name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "422": {
                        "description": "Unproccessed data"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/tag": {
            "post": {
                "tags": [
                    "Tag"
                ],
                "summary": "Creates and returns a tag object",
                "description": "Creates and returns a tag",
                "operationId": "createTag",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Tag name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sucess"
                    },
                    "422": {
                        "description": "Unproccessed data"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        }
    }
}